# [Xamarin.Android.Build.Tasks] no temp files in GenerateJavaStubs