--- a/src/Xamarin.Android.Build.Tasks/Utilities/Files.cs
+++ b/src/Xamarin.Android.Build.Tasks/Utilities/Files.cs
@@ -4,6 +4,7 @@ using System.Security.Cryptography;
 
 using Xamarin.Tools.Zip;
 using System.Collections.Generic;
+using System.Text;
 #if MSBUILD
 using Microsoft.Build.Utilities;
 using Xamarin.Android.Tasks;
@@ -69,6 +70,44 @@ namespace Xamarin.Android.Tools {
 			}/* else
 				Console.WriteLine ("Skipping copying {0}, unchanged", Path.GetFileName (destination));*/
 
+			return false;
+		}
+
+		public static bool CopyIfStringChanged (string contents, string destination)
+		{
+			//NOTE: this is not optimal since it allocates a byte[]. We can improve this down the road with Span<T> or System.Buffers.
+			var bytes = Encoding.UTF8.GetBytes (contents);
+			return CopyIfBytesChanged (bytes, destination);
+		}
+
+		public static bool CopyIfBytesChanged (byte[] bytes, string destination)
+		{
+			if (HasBytesChanged (bytes, destination)) {
+				var directory = Path.GetDirectoryName (destination);
+				if (!string.IsNullOrEmpty (directory))
+					Directory.CreateDirectory (directory);
+
+				MonoAndroidHelper.SetWriteable (destination);
+				File.WriteAllBytes (destination, bytes);
+				return true;
+			}
+			return false;
+		}
+
+		public static bool CopyIfStreamChanged (Stream stream, string destination)
+		{
+			if (HasStreamChanged (stream, destination)) {
+				var directory = Path.GetDirectoryName (destination);
+				if (!string.IsNullOrEmpty (directory))
+					Directory.CreateDirectory (directory);
+
+				MonoAndroidHelper.SetWriteable (destination);
+				using (var fileStream = File.Create (destination)) {
+					stream.Position = 0; //HasStreamChanged read to the end
+					stream.CopyTo (fileStream);
+				}
+				return true;
+			}
 			return false;
 		}
 
@@ -162,7 +201,39 @@ namespace Xamarin.Android.Tools {
 			var src_hash = HashFile (source);
 			var dst_hash = HashFile (destination);
 
-			// If the hashed don't match, then the file has changed
+			// If the hashes don't match, then the file has changed
+			if (src_hash != dst_hash)
+				return true;
+
+			return false;
+		}
+
+		public static bool HasStreamChanged (Stream source, string destination)
+		{
+			//If destination is missing, that's definitely a change
+			if (!File.Exists (destination))
+				return true;
+
+			var src_hash = HashStream (source);
+			var dst_hash = HashFile (destination);
+
+			// If the hashes don't match, then the file has changed
+			if (src_hash != dst_hash)
+				return true;
+
+			return false;
+		}
+
+		public static bool HasBytesChanged (byte [] bytes, string destination)
+		{
+			//If destination is missing, that's definitely a change
+			if (!File.Exists (destination))
+				return true;
+
+			var src_hash = HashBytes (bytes);
+			var dst_hash = HashFile (destination);
+
+			// If the hashes don't match, then the file has changed
 			if (src_hash != dst_hash)
 				return true;
 
@@ -272,8 +343,15 @@ namespace Xamarin.Android.Tools {
 
 		public static string HashString (string s)
 		{
+			var bytes = Encoding.UTF8.GetBytes (s);
+			return HashBytes (bytes);
+		}
+
+		public static string HashBytes (byte [] bytes)
+		{
 			using (HashAlgorithm hashAlg = new SHA1Managed ()) {
-				return HashFile (s, hashAlg);
+				byte [] hash = hashAlg.ComputeHash (bytes);
+				return BitConverter.ToString (hash);
 			}
 		}
 
@@ -288,13 +366,14 @@ namespace Xamarin.Android.Tools {
 		{
 			using (Stream file = new FileStream (filename, FileMode.Open, FileAccess.Read)) {
 				byte[] hash = hashAlg.ComputeHash (file);
-
 				return BitConverter.ToString (hash);
 			}
 		}
 
 		public static string HashStream (Stream stream)
 		{
+			stream.Position = 0;
+
 			using (HashAlgorithm hashAlg = new SHA1Managed ()) {
 				byte[] hash = hashAlg.ComputeHash (stream);
 				return BitConverter.ToString (hash);