--- a/src/Xamarin.Android.Build.Tasks/Utilities/ManifestDocument.cs
+++ b/src/Xamarin.Android.Build.Tasks/Utilities/ManifestDocument.cs
@@ -928,7 +928,13 @@ namespace Xamarin.Android.Tasks {
 		
 		public void Save (string filename)
 		{
-			using (var file = new StreamWriter (filename, false, new UTF8Encoding (false)))
+			using (var file = new StreamWriter (filename, append: false, encoding: new UTF8Encoding (false)))
+				Save (file);
+		}
+
+		public void Save (Stream stream)
+		{
+			using (var file = new StreamWriter (stream, new UTF8Encoding (false), bufferSize: 1024, leaveOpen: true))
 				Save (file);
 		}