@jonathanpeppers 2018-12-14
Logs from third run: [logs.zip](https://github.com/xamarin/xamarin-android/files/2678965/logs.zip)