@jonathanpeppers 2018-12-18
[The test failure](https://jenkins.mono-project.com/job/xamarin-android-pr-builder-release/222/testReport/junit/_Users_builder_jenkins_workspace_xamarin-android-pr-builder-release_xamarin-android_build-tools_scripts_.._.._TestResult-Xamarin.Android.Bcl_Tests.nunit-Release/xml/TestResult_Xamarin_Android_Bcl_Tests_nunit_Release_xml/) just seems to say this from the build log:
```
/Users/builder/jenkins/workspace/xamarin-android-pr-builder-release/xamarin-android/build-tools/scripts/TestApks.targets(189,5):   Process '/Users/builder/android-toolchain/sdk/platform-tools/adb -s emulator-5570  shell am instrument  -e "loglevel Verbose" -w "Xamarin.Android.Bcl_Tests/xamarin.android.bcltests.NUnitInstrumentation"' failed to exit within the timeout of 1800000ms, killing the process [/Users/builder/jenkins/workspace/xamarin-android-pr-builder-release/xamarin-android/tests/RunApkTests.targets]
```
I'm going to retry it.